+++ demo/RandomForest.R
@@ -13,7 +13,7 @@ test  <- wine[-train.index,]
 cls <- which(colnames(wine) == "Wine")
 
 #% LABELED
-labeled.index <- createDataPartition(wine$Wine, p = .2, list = FALSE)
+labeled.index <- createDataPartition(train$Wine, p = .2, list = FALSE)
 train[-labeled.index,cls] <- NA
 
 
@@ -27,3 +27,4 @@ predict(m,test) %>%
 
 #For probabilities
 predict(m,test, type = "prob")
+