+++ inst/doc/regression.R
@@ -1,9 +1,9 @@
-## ----eval=FALSE---------------------------------------------------------------
+## ----eval=FALSE, warning=FALSE,message=FALSE----------------------------------
 #  
 #  library(SSLR)
 #  library(tidymodels)
 
-## ----include=FALSE------------------------------------------------------------
+## ----libraries, results="hide", warning=FALSE,message=FALSE-------------------
 knitr::opts_chunk$set(
   digits = 3,
   collapse = TRUE,
@@ -18,13 +18,14 @@ library(tidymodels)
 set.seed(1)
 
 data <- airquality
-
+#Delete column Solar.R (NAs values)
+data$Solar.R <- NULL
 #Train and test data
 train.index  <- sample(nrow(data), round(0.7 * nrow(data)))
 train <- data[ train.index,]
 test  <- data[-train.index,]
 
-cls <- which(colnames(wine) == "Ozone")
+cls <- which(colnames(airquality) == "Ozone")
 
 #% LABELED
 labeled.index <- sample(nrow(train), round(0.1 * nrow(train)))
@@ -39,9 +40,17 @@ predict(m,test)%>%
   bind_cols(test) %>%
   metrics(truth = "Ozone", estimate = .pred)
 
+## ----fitrf, results="hide"----------------------------------------------------
+m <- SSLRRandomForest(trees = 5,  w = 0.3) %>% fit(Ozone ~ ., data = train)
+
 ## ----fitcobc, results="hide", eval = FALSE------------------------------------
 #  m_r <- rand_forest( mode = "regression") %>%
 #    set_engine("ranger")
 #  
-#  m <- coBC(learner = m_r, max.iter = 2) %>% fit(Ozone ~ ., data = train)
+#  m <- coBC(learner = m_r, max.iter = 1) %>% fit(Ozone ~ ., data = train)
+
+## ----fitcoreg, results="hide", eval = FALSE-----------------------------------
+#  #Load kknn
+#  library(kknn)
+#  m_coreg <- COREG(max.iter = 1)  %>% fit(Ozone ~ ., data = train)